+++ History.md
@@ -190,4 +190,4 @@
  * + Removed the trailing string 'exception' from event names
  * + Renamed 'abortedexception' ( 'aborted' ) to 'abort'
  * + Renamed 'end' to 'loadend'
-
+ * + Renamed 'dataprogress' to 'progress'